feat: added optional prefix/suffix to do advanced matching
An optional prefix/suffix was added to support files that have other
data in them. This way, the parsing does not start until the prefix is
found, after which it doesn't stop until the suffix is found. This
simplifies the need to do advanced pattern matching, and allows any
number of blank lines to be included (useful when prefilling a timesheet
for a day and adding times for meetings, but you want to have a visual
indication of where the gaps exist.